read control temperature, add date type adapter

+++ app/src/main/java/com/ashtonandassociates/thermopi/api/shared/ApiTemperature.java
@@ -8,10 +8,10 @@ public class ApiTemperature {
 	public static final int CONST_API_SCALE = 0; // mal 10000
 	public static final int CONST_DEFAULT_SCALE = 1; // normal
 
-	private Double temperature;
+	private Integer temperature;
 	private int scale = CONST_DEFAULT_SCALE;
 
-	public ApiTemperature(Double temperature, int scale) {
+	public ApiTemperature(Integer temperature, int scale) {
 		this.temperature = temperature;
 		this.scale = scale;
 	}
@@ -21,19 +21,19 @@ public class ApiTemperature {
 	 * @param param
 	 * @param scale
 	 */
-	public void setTemperature(Double param, int scale) {
+	public void setTemperature(Integer param, int scale) {
 		this.scale = scale;
 		this.temperature = param;
 	}
 
 	public Integer getTemperature(int scale) {
-		Double retVal = null;
+		Integer retVal = null;
 		if(this.scale == scale) {
 			retVal = this.temperature;
 		} else if(this.scale == CONST_DEFAULT_SCALE && scale == CONST_API_SCALE) {
-			retVal = this.temperature.doubleValue() * 1000;
+			retVal = this.temperature.intValue() * 1000;
 		} else if(this.scale == CONST_API_SCALE && scale == CONST_DEFAULT_SCALE) {
-			retVal = this.temperature.doubleValue() / 1000;
+			retVal = this.temperature.intValue() / 1000;
 		}
 		return retVal.intValue();
 	}
